apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "filesentinel-api"
version = "0.1.0"
description = "FastAPI service for FileSentinel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.12",
  "uvicorn[standard]>=0.34.2",
  "magika>=0.6.2",
  "python-multipart>=0.0.20",
  "python-jose[cryptography]>=3.3.0",
  "pydantic-settings>=2.9.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "httpx>=0.28.1"
]
